requirements.txt
selenium
undetected-chromedriver
supabase
python-dotenv
webdriver-manager
setuptools